Add sub mod memory for mm
